[RED] WIP DiscreteAnnealType Tests Compile
I really want to change the type of the state/state_curr etc. variables
back to integer from real. I had to change them to real in order to make
them compatible with operations with sparse_blas/rsb, but they really
need to be interpreted as integers. (Especially for the Discrete
Annealing case.)