[RED] WIP on Changing State Type
In order to get this stuff to compile, I had to change the state
variable to the real32 type (so that it had a match within the BLAS
standard). But I'd like to change it to an integer type, since that is
what it is for the discrete anneal tests anyways. But there is an issue
rn with missing type members.