[GREEN] Hardcoded Ising Model Config into Tests
Now the discrete anneal test has one hardcoded solution built into it.
Potentially in the future, more cleanup should be done on the
ising_test_setup.sage file to simply print to STDOUT the values that
were hardcoded, or the test should be made to read from the datafile
that the sage script produces. Although this would introduce further
test randomness.